Enable fetching role descriptions
Role session durations are not returned by the GetAccountAuthorizationDetails
call, they are fetched later in the populateIamData function if
SkipFetchingPolicyAndRoleDescriptions is true.

This is a quick fix, given we always (both pull and push) need to fetch the
MaxSessionDuration attribute we should really remove all of the plumbing for
the skipping.