Fetch Policy Tags
We want to be able to ignore policies as well as roles using
our standard tagging mechanisms.  Tagging of policies is a
relatively new feature, but it is in general release.  We need
it to support programmatically generated things like the
roles and policies that Cloudwatch dashboards generates, which
add a random label at the end of the resource name, but a differently
formatted label than cloudwatch uses :-(